Remove automatic subnet building
Before this stacker would automatically determine availability zones and
both public & private subnets, based on subnet math. It put too much of
the burden on python. Now, instead, the AZs are determined
automatically, but in CloudFormation (Fn::GetAZs) and you only provide a
count of AZs.  As well, you need to provide a set of Public & Private
subnets to use.