Pass a new context to each blueprint
The first version of stacker passed a new context object to each
blueprint. Since we're going to stop passing down the context to the
blueprint in the future, we only create one Context.

This fix is temporary until we finalize what should actually be passed
to the blueprint, ie. only namespace and parameters.