Address comments, clean up extra params
This tries to pull in every DBInstance Parameter in as a 'default'.  The
only ones not modified are:

AvailabilityZone: Can't be used w/ MultiAZ, can be added later or in
                  subclasses
CharacterSetName: Oracle only
DBInstanceIdentifier: Very dependent on the underlying Engine
DBSecurityGroups: Not used in VPC
DBSnapshotIdentifier: Used to create database from backup, could be done
                      later
Port: Default dependent on Engine, don't want to mess with it.
PubliclyAccessible: Kind of hard to use, not the primary use case really