Remove storage of stacks on context object
Fixes GH-111

This will be, in theory, less efficient since every time you call
get_stacks it will have to rebuild the Stack objets, but in the end it
saves the nasty memory issue we have, and it's simple.

If it ever becomes a problem we could store the stack objects in a
STACKS global or something similar, but that seems gross, and the type
of pre-optimization that'll bite you in the ass (like this did).