Change params from None to blank string
There's (almost?) never a good reason to pass 'None' as a Parameter to a
cfn template, so in the case that None is passed in (usually due to a
blank parameter in the config/env), change it to a blank string. I think
this is almost always what is expected.

This also makes a function called resolve_parameters, pulling it out of
the build.Action class, so it is easier to test.