Blank parameters now result in removing the param
I think this actually makes more sense - it allows us to fall back to
defaults for Parameters a lot easier by just not specifying a value.
Ie:

Param:

^^ Nothing after the Param, which is valid yaml, results in that Param
not being submitted to the blueprint, which means it'll fall back on
it's default.  If we want blank strings, people should be explicit about
that, ie:

Param: !!str
or
Param: ""