Convert python booleans in parameters to strings
Fixes GH-135

Some resources have the concept of boolean Parameters, but these are
booleans in json, which is a lowercase string.  This allows people to
use `true` or `false` in boolean parameters, without having to specify
`!!str` before it in order to force the yaml to not turn them into
python booleans.