1.0: Stop people from accidentally using old parameters
I ran into an issue while testing the change from Parameters to
Variables while updating the firehose blueprint in stacker_blueprints.
All the variables are optional, which means if you don't provide any
of them, it'll happily move along.  This is bad, since say before you
had a parameter for `BucketName`.  Now you haven't provided one for
the Variable `BucketName`, so CF assumes it should create a new bucket
with a random name, and then it will delete your old bucket.

Since Parameters, as they stood, no longer exist in 1.0, I think this
extra level of protection is a good idea.  I might be missing somewhere
else that we should protect people as well, so any ideas on that would
be greatly appreciated.