Drop implicit list syntax
We now use `TroposphereType(s3.Bucket, many=True)` to indicate that a
type will create more that one tropopshere type. Since the value
provided in the config is always a dictionary, this is easier to
understand than the previous `TroposphereType([s3.Bucket])` syntax,
since you're not passing in a list.