Remove --var flag
--variables comes from the days when we didn't have environments, so the
only way to change things between environments was to pass arguments on
the command line.  In general, environments are a better use case
(reproducibility is much better) and continuing to support the CLI --var
was getting difficult due to complex variable types.

Removing for 1.0 after discussing on #stacker. If needed in the future,
we can reconsider.