Initial spike @ using click for the CLI
Our current Commands setup is crazy hard to follow, with crazy
inheritance, and I'm not sure if it really gives us anything. Rather
than re-create the wheel, I decided to give the click library a try.

Not sure I like it all that much more. Coming up with "common"
parameters is a bit of a mess. Figured I'd open this up to at least get
some comments around.