Update hooks to use the provider for aws interactions
Rather than setting up their own sessions, we already pass the provider
in.  By standardizing on using the provider for all interactions, we can
share logic between all the connections, which I think is beneficial.

In doing this I discovered that the mock_provider really wasn't
necessary any longer, and we could use an actual Provider class for all
our tests.

This is all in preparation for trying to get to #374, and make it less
likely that we'll regress once again.