Add support for Property objects in TroposphereType variables
Previously TroposphereType only worked for complete resources, not
property objects meant to be used as parameters for other resources.

Change that by inspecting the type to determine if it represents a property
object, and interpreting the variable value differently. A dictionary is
used directly as parameters, instead as title/parameters pairs, and a list
can be provided to generate multiple objects (if the type definition
allows).

Also add an `optional` boolean parameter to the constructor (mutually
exclusive to `many`), indicating that a single object should be created,
but that a `None` (or ommited) value is acceptable, in which case it is
passed through to the template.