new bucket region config
Fixes #413. Adds a new config setting, `stacker_bucket_region`. If set
the stacker bucket where templates are stored is created in that region.
If not, it defaults to using the region provided by `--region` or
`AWS_REGION`.

This is a semi-breaking change, unfortunately - simply because we
weren't consistent about this before.