Always use us-east-1 with s3 connections
Turns out you can do anything in any region with s3 provided your
connection is in us-east-1 (global).  That's not true of other regions -
if you connect to us-west-1 for example, you can only work in us-west-1.

This is the "easiest" way to work with this.  There might be some
compliance concerns for more strict places (US Gov probably?), but I'm
not ready to dig into that yet - if someone else is concerned, all for a
patch :)