testutil: make blueprint result comparison more robust and easy to read
Compare results of blueprint tests to the expected fixtures directly as
dictionaries, by using `to_dict`, and parsing the file with
`json.loads`, respectively. Then generate the strings with `json.dumps`
with the same parameters (with `sort_keys` enabled) such that
differences in the output are easy to discerne.

This should make false-negatives much less likely.