Add support for automatically re-creating stacks in un-updatable states
During initial development of blueprints and/or configs, it's common for
stacks to fail validation. When that happens, they are left in the
ROLLBACK_COMPLETE or CREATE_FAILED states, in which it is not possible
to make any updates. The user must manually delete the stack through the
AWS CLI or Console, then try again.

If we know that the state is a safe state to delete, that it was created
by stacker (by observing it's tags), and that we have user confirmation
through a prompt or an explicit command line option, we can just delete
the stack and re-create it without manual intervention.

Implement support for that workflow by verifying if a stack is updatable
before trying, and if negative, doing a delete-then-create sequence
instead.