Update Plan to wait for rollbacks to finish instead of panicking
There's no good reason why we cannot re-use the logic to wait for
submitted stacks to wait for their own rollbacks. Now that we handle
re-creating failed stacks, it's much better if stacker always wait until
there is nothing else in progress, so that the next call can fix things
up.

To achieve that, pay attention to rollback states in the default
provider, and introduce a FAILED state to steps. Steps fail if their
run_func fails, or if any of the dependencies fail before it runs.

If any step of the whole plan fails, and exception is raised in the end.