Add handling of stack rollbacks and recreations (#473)
* Add support for automatically re-creating stacks in un-updatable states

During initial development of blueprints and/or configs, it's common for
stacks to fail validation. When that happens, they are left in the
ROLLBACK_COMPLETE or CREATE_FAILED states, in which it is not possible
to make any updates. The user must manually delete the stack through the
AWS CLI or Console, then try again.

If we know that the state is a safe state to delete, that it was created
by stacker (by observing it's tags), and that we have user confirmation
through a prompt or an explicit command line option, we can just delete
the stack and re-create it without manual intervention.

Implement support for that workflow by verifying if a stack is updatable
before trying, and if negative, doing a delete-then-create sequence
instead.

* Add --recreate-failed option to CLI

* Update integration tests to generate access key automatically

* Update Plan to wait for rollbacks to finish instead of panicking

There's no good reason why we cannot re-use the logic to wait for
submitted stacks to wait for their own rollbacks. Now that we handle
re-creating failed stacks, it's much better if stacker always wait until
there is nothing else in progress, so that the next call can fix things
up.

To achieve that, pay attention to rollback states in the default
provider, and introduce a FAILED state to steps. Steps fail if their
run_func fails, or if any of the dependencies fail before it runs.

If any step of the whole plan fails, and exception is raised in the end.

* Update build tests to better verify the different cases

Split up each one into a different end-to-end test, and use actual state
names instead of mocking every single checking method.

* Add stack re-creation integration tests

* Add update rollback integration tests

* Make it possible for a plan to fail gracefully without printing a TB

* Improve error message when a stack fails due to a dependency

* providers/aws: rename check_tags_present -> check_tags_contain

And add docstring.

* Document handling of UPDATE_ROLLBACK_COMPLETE stack state

Since it signals both completion and failure, it needs to be handlec
with a bit more care than the other states.

* Improve stack re-creation user messages

- Point out how to enable re-creation when it is disabled (as
suggested by @GarisonLotus)
- Be a little more verbose when a stack is in an invalid state for
re-creation
- Remove redundant log message when destroying a stack for re-creation

* tests: only enable recreate_failed in provider when needed

* Remove reference to 'plan' when printing failed stacks

* tests: document how the broken mock blueprint actually breaks

The `Timeout` parameter is made excessively large to make AWS reject
it.

* docs: update commands with new --recreate-failed option

* exceptions: remove unnecessary parentheses