Move S3 templates into sub-dirs in S3 (#478)
* Move S3 templates into sub-dirs in S3

Fixes #347

This adds another layer to the s3 bucket space to try and keep things a
bit cleaner.  If you had a namespace of `mynamespace`, a bucket of
`mybucket`, and a stack called `vpc`, the old method would have resulted
in an s3 key of:

`s3://mybucket/vpc-$version.json`

It would now result in:

`s3://mybucket/mynamespace-vpc/vpc-$version.json`

Not 100% sure that's the best namespace use, but it seems specific
enough and should be helpful.

I'm pretty sure this will result in slower runs the first time after
updating to this version of stacker, only because it'll have to upload
all it's templates again, but it shouldn't result in any CloudFormation
updates.  I'll verify this.

* Prepend s3 path with `/stack_templates/`