Remove changesets if no changes were made
Fixes #508

Not sure when this was introduced, but there's no reason to leave behind
ChangeSets that were created but not used because no changes actually
happened.

We still leave behind changesets that have an abnormal status for
investigations, though maybe in the future this should be an argument or
something.