Always dump the current output/result (#528)
* Always dump the current output/result

This fixes an issue that was introduced in #467. The original
functionality was to always dump the current output to `.json-result`,
that way it was easy to build & compare the output.

With the current implementation (from #467) we raise an exception if the
old results aren't available, but don't create the new results.  This is
an issue whenever creating a new test.

* Update CHANGELOG