Add support for "targets"
This adds "targets" to stacker, which are just named nodes in the graph
that specify dependencies.

Why would you want this? Primarily just as an easy way to specify a
logical grouping of stacks that you can use with the --targets flag
(e.g. maybe you want a target for "databases" or "apps", etc).
Essentially, these are no different than stacks, except they don't incur
any AWS API calls, and don't take up a precious CloudFormation stack
spot.

If you're familiar with targets in systemd, the concept is exactly the
same; a node in the graph that does nothing.