Change the way we validate graph output.
Key order has changed between python2 and python3, so this old test that
relies on python2 key order behavior is no longer working. For a
comparison, this is the output of the command in Python 3:

```
{
    "steps": {
        "vpc": {
            "deps": []
        },
        "bastion2": {
            "deps": [
                "vpc"
            ]
        },
        "bastion1": {
            "deps": [
                "vpc"
            ]
        },
        "app2": {
            "deps": [
                "bastion2"
            ]
        },
        "app1": {
            "deps": [
                "bastion1"
            ]
        }
    }
}
```

which, apart from the key order, is the same with the output embeded in
the results.

In order to support Python 2 and Python 3 at the same time, I have
validated three keys in that json output, hopes that's enough.

Signed-off-by: Kai Xia <xiaket@gmail.com>