Hack PyYAML so it will use OrderedDict for mapping.
This function aiming to keep the order from the yaml file in an OrderedDict
in the output of the yaml.load output, as well as make sure that some of the
keys are unique.

Our previous attempt to keep the order in the yaml in an OrderedDict is
flawed. Take our test case for example:

```
raw_config = """
pre_build:
  hook2:
    path: foo.bar
  hook1:
    path: foo1.bar1
"""
config = yaml_to_ordered_dict(raw_config)
```

The type of the config instance is dict instead of OrderedDict, moreover,
the type of config['pre_build'] is also a dict. We fixed this behavior
in this commit.

PyYAML used dict internally to save mappings. This hardcoding of dict happens
in two places, one in `SafeLoader.construct_yaml_map`, the other happens in our
`_validate_mapping`, where a dictionary is used to keep the validated keys.
This validated mapping is then sent to OrderedDict, but at this stage,
all the key order is scrambled.

So we need to:

1. Override the default behavior of `SafeLoader.construct_yaml_map`.
2. Use OrderedDict in `_validate_mapping`, and return it directly.

The second one is easy to change, however, we have two choices to override
the default method in `SafeLoader`. We can either monkey patch it:

```
yaml.loaders.SafeLoader.construct_yaml_map = our_construct_yaml_map
```

Or, the way I've used in this commit:

```
OrderedUniqueLoader.add_constructor(
    u'tag:yaml.org,2002:map', OrderedUniqueLoader.construct_yaml_map,
)
```

The first method is by far the cleaner solution, except that it's a
monkey patch which may in some unforeseeable way bring disaster to us.
The current solutions involves the hardcoding of the type identifier,
which normally is not good, but even the PyYAML devs are hardcoding this
string in 3 different places so I think it's safe.

Signed-off-by: Kai Xia <xiaket@gmail.com>