fix indeterministic result in DAG.topological_sort.
In the `topological_sort` method of `DAG` class, `in_degree` is a dictionary
to keep track of how deep an element is dependent: the higher the value, the
lower the element in the dependence tree.

In the test case, the initialization dict is:

```
{
    'a': ['b', 'c'],
    'b': ['d'],
    'c': ['d'],
    'd': [],
}
```

The calculated `in_degree` is:

```
{'a': 0, 'b': 1, 'c': 1, 'd': 2}
```

which looks good so far. Then later in the while loop, we are getting the
elements of the same "in degree" from `graph[u]`, which in fact is a set and
the ordering of the elements are in general indeterministic. Our test case
happen to have (again) relied on that specific order, which is changed in
Python 3. That's why half of the time we can see a failure from this test.

The solution is simple, since we are facing elements of the same "in degree"
here, we could simply order the elements by their name by adding a `sorted()`
call around `graph[u]`, which will return a deterministic list of elements.

Signed-off-by: Kai Xia <xiaket@gmail.com>