Change test runner from nose to py.test  (#714)
* Change test runner from nose to py.test

Nose has been deprecated by its authors. py.test is very active, widely
used, and has many useful features for running and writing tests.

Changing the tests themselves away from unittest-style is a different
step that must be taken later. This commit only changes the tooling.

* dag: port tests to pytest

* blueprints: raw: always return absolute template paths

Additionaly, adapt tests to not depend on any fixtures in the test
directory, or on the current working directory of the test runner.

* test: disable deprecation warnings by default due to noise