Add YAML environment file support
Environment files ending in .yaml or .yml will be rendered using
a different strategy. The env file is treated as a dict, with
the value for each key being a parsed YAML object. After parsing
the config, we look for things that look like variable references,
and replace the references with objects.