Add YAML environment file support (#740)
* Add YAML environment file support

Environment files ending in .yaml or .yml will be rendered using
a different strategy. The env file is treated as a dict, with
the value for each key being a parsed YAML object. After parsing
the config, we look for things that look like variable references,
and replace the references with objects.

* Fix lint errors

* Fix more lint errors

Now with the correct version of flake8

* Fix IAM test errors due to missing version

Moto is requiring policy documents to be at least version
2012-10-17, so add that property.

* Fix python 2.7 issues

Python 2.7 strings and Python 3 strings work differently, and
regular expression matching is different too. Rewrite in a way
which works with both language versions.

* Remove debug printout

* Fix IAM test errors due to missing version (#741)

Moto is requiring policy documents to be at least version
2012-10-17, so add that property.

* Use case insensitive file match

When checking for yaml suffixes on environment files, ignore
case.

* Add YAML environment file support

Environment files ending in .yaml or .yml will be rendered using
a different strategy. The env file is treated as a dict, with
the value for each key being a parsed YAML object. After parsing
the config, we look for things that look like variable references,
and replace the references with objects.

* Fix lint errors

* Fix more lint errors

Now with the correct version of flake8

* Fix python 2.7 issues

Python 2.7 strings and Python 3 strings work differently, and
regular expression matching is different too. Rewrite in a way
which works with both language versions.

* Remove debug printout

* Use case insensitive file match

When checking for yaml suffixes on environment files, ignore
case.

* Small documentation update

Yaml config example should use .yml file suffix, not .env

Co-authored-by: Marcin Romaszewicz <47459980+deepmap-marcinr@users.noreply.github.com>